+++ mariner/mariner_test.go
@@ -1,4 +1,4 @@
-package cblmariner_test
+package mariner_test
 
 import (
 	"io/fs"
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	cblmariner "github.com/aquasecurity/vuln-list-update/cbl-mariner"
+	mariner "github.com/aquasecurity/vuln-list-update/mariner"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -25,14 +25,14 @@ func TestUpdate(t *testing.T) {
 		{
 			name:      "sad path, invalid xml",
 			inputFile: "file::testdata/sad/invalid",
-			wantError: "failed to update oval data: failed to unmarshal xml: XML syntax error on line 80: unexpected EOF",
+			wantError: "failed to update oval data: failed to decode xml: XML syntax error on line 80: unexpected EOF",
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			tmpDir := t.TempDir()
-			cc := cblmariner.NewConfig(cblmariner.WithURLs(tt.inputFile), cblmariner.WithDir(tmpDir), cblmariner.WithRetry(0))
+			cc := mariner.NewConfig(mariner.WithURLs(tt.inputFile), mariner.WithDir(tmpDir), mariner.WithRetry(0))
 			if err := cc.Update(); tt.wantError != "" {
 				require.Error(t, err)
 				assert.Contains(t, err.Error(), tt.wantError)
@@ -49,7 +49,7 @@ func TestUpdate(t *testing.T) {
 				got, err := os.ReadFile(filepath.Join(tmpDir, path))
 				assert.NoError(t, err, tt.name)
 
-				goldenPath := filepath.Join("testdata", "golden", "cbl-mariner", path)
+				goldenPath := filepath.Join("testdata", "golden", "mariner", path)
 				want, err := os.ReadFile(goldenPath)
 				assert.NoError(t, err, tt.name)
