+++ mariner/testdata/golden/mariner/1.0/definitions/2008/CVE-2008-3914.json
