+++ mariner/testdata/golden/mariner/1.0/definitions/2021/CVE-2021-35942.json
