+++ mariner/testdata/golden/mariner/1.0/objects/objects.json
