+++ mariner/testdata/golden/mariner/2.0/definitions/2014/CVE-2014-8139.json
