+++ mariner/testdata/golden/mariner/2.0/definitions/2021/CVE-2021-39924.json
