+++ mariner/testdata/golden/mariner/2.0/definitions/2022/CVE-2022-21309.json
