+++ mariner/testdata/golden/mariner/2.0/objects/objects.json
